copilot refactor of mocking aws rum
